Update semantics.md to improve readability of example
In my experience an example is more readable if additional detail is introduced after the main fact. In this example the property :b is part of the main fact and the owl:sameAs relation between :b and :c is the additional detail. First seeing :b used and then being informed about the relation of :b to something else makes the example more readable IMHO.