Update semantics.md to improve example understanding
In my own and painful experience some small hints can make it much easier for non-logicians to understand what is going on and what an example tries to convey. 
The concrete changes I propose may not be fully correct or grasp what's going to the fullest extend, so take them as a prototype to illustrate the issue.
I deliberately used the term 'means' instead of 'denotes' or 'refers to' beacuse it is more approachable, but the latter might be more suitable as they are used much more frequently in the specs.