Update semantics.md to better describe the meaning of opacity
The document should explain in greater detail to non-logicians (which I assume to be its main target audience, right?) what it means that the syntax of a term is important. It doesn't mean that the term is just a string (right?!) but that its interpretation is restricted. This pull request tries to explain it, and although it probably fails some explanation IMO is necessary.