add document on semantics options and add options to terminology document (#66)
* add document on semantics options and add options to terminology document

* fix link syntax

* Update docs/semantics.md

Co-authored-by: Ted Thibodeau Jr <tthibodeau@openlinksw.com>

* incorporate suggestions

* Update docs/semantics.md

Co-authored-by: Ted Thibodeau Jr <tthibodeau@openlinksw.com>

* Update docs/rdf-terminology.md

Co-authored-by: Ted Thibodeau Jr <tthibodeau@openlinksw.com>

* Update semantics.md

Keep the use of xsd:int for the forth variant consistent across different examples

* Update semantics.md to improve example understanding

In my own and painful experience some small hints can make it much easier for non-logicians to understand what is going on and what an example tries to convey. 
The concrete changes I propose may not be fully correct or grasp what's going to the fullest extend, so take them as a prototype to illustrate the issue.
I deliberately used the term 'means' instead of 'denotes' or 'refers to' beacuse it is more approachable, but the latter might be more suitable as they are used much more frequently in the specs.

* Update semantics.md to better describe the meaning of opacity

The document should explain in greater detail to non-logicians (which I assume to be its main target audience, right?) what it means that the syntax of a term is important. It doesn't mean that the term is just a string (right?!) but that its interpretation is restricted. This pull request tries to explain it, and although it probably fails some explanation IMO is necessary.

* Update semantics.md

fixed typo

---------

Co-authored-by: Adrian Gschwend <ktk@netlabs.org>
Co-authored-by: Ted Thibodeau Jr <tthibodeau@openlinksw.com>
Co-authored-by: Thomas Lörtsch <tl@rat.io>